js to ts implementation pages/Chat
