+++ src/components/Checkbox/index.tsx
@@ -2,20 +2,19 @@ import Icon from "components/Icon";
 import React from "react";
 import "./styles.css";
 
-const Checkbox = ({ inputProps }) => {
-	return (
-		<div className="checkbox">
-			<input
-				type="checkbox"
-				id="checkbox"
-				{...inputProps}
-				className="checkbox__input"
-			/>
-			<label htmlFor="checkbox" className="checkbox__label">
-				<Icon id="check" className="checkbox__icon" />
-			</label>
-		</div>
-	);
+type Props = {
+  [x: string]: any;
+};
+
+const Checkbox = ({ inputProps }: Props) => {
+  return (
+    <div className="checkbox">
+      <input type="checkbox" id="checkbox" {...inputProps} className="checkbox__input" />
+      <label htmlFor="checkbox" className="checkbox__label">
+        <Icon id="check" className="checkbox__icon" />
+      </label>
+    </div>
+  );
 };
 
 export default Checkbox;
